Crate variadics

Crate for macro-free variadic tuple metaprogramming.

Rationale

As of writing this crate, Rust does not support variadic generics and does not allow to reason about tuples in general.

Most importantly, Rust does not allow one to generically implement a trait for all tuples whose elements implement it.

This crate attempts to fill the gap by providing a way to recursively define traits for tuples.

Tuple lists

Tuple (A, B, C, D) can be unambiguously mapped into recursive tuple (A, (B, (C, (D, ())))).

On each level it consists of a pair (Head, Tail), where Head is tuple element and Tail is a remainder of the list. For last element Tail is an empty list.

Unlike regular flat tuples, such recursive tuples can be effectively reasoned about in Rust.

This crate calls such structures “variadics” and provides a set of traits and macros allowing one to conveniently work with them.

Acknowledgements

This crate is based on tuple_list by VFLashM

MIT License

Copyright (c) 2020 Valerii Lashmanov

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
SOFTWARE.

Macros

var_args

Macro for pattern-matching with variadic tuples. This is mainly used for function arguments, but it can also be used in match, if let ..., and let ... else expressions.

var_expr

Macro creating a variadic tuple value from a list of expressions.

var_type

Macro creating a variadic tuple type from a list of types.

variadic_trait

This macro generates a basic variadic trait where each element must fulfill the trait clause. Currently of limited usefulness.

Traits

Extend

Split

SplitPrefix

Variadic