variadics

Trait SplitBySuffix

pub trait SplitBySuffix<Suffix>: VariadicExt + Sealed<Suffix>
where
    Suffix: VariadicExt,
{
    type Prefix: VariadicExt;

    // Required methods
    fn split_by_suffix(self) -> (Self::Prefix, Suffix);
    fn split_by_suffix_ref(
        this: Self::AsRefVar<'_>,
    ) -> (<Self::Prefix as VariadicExt>::AsRefVar<'_>, Suffix::AsRefVar<'_>);
}

Helper trait for splitting a variadic into two parts. Prefix is the first part, everything after is the Suffix or second part.

This is a sealed trait.

Required Associated Types

type Prefix: VariadicExt

The first part when splitting this variadic by Suffix.

Required Methods

fn split_by_suffix(self) -> (Self::Prefix, Suffix)

Splits this variadic into two parts, first the Prefix, and second the Suffix.

fn split_by_suffix_ref( this: Self::AsRefVar<'_>, ) -> (<Self::Prefix as VariadicExt>::AsRefVar<'_>, Suffix::AsRefVar<'_>)

Splits a refvar variadic

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.