Trait variadics::VariadicExt

pub trait VariadicExt: Variadic + Sealed {
    type Extend<Suffix>: VariadicExt
       where Suffix: VariadicExt;
    type Reverse: VariadicExt;
    type AsRefVar<'a>: RefVariadic<UnRefVar = Self, RefVar = Self::AsRefVar<'a>, MutVar = Self::AsMutVar<'a>>
       where Self: 'a;
    type AsMutVar<'a>: MutVariadic<UnRefVar = Self, RefVar = Self::AsRefVar<'a>, MutVar = Self::AsMutVar<'a>>
       where Self: 'a;
    type IterAnyRef<'a>: Iterator<Item = &'a dyn Any>
       where Self: 'static;
    type IterAnyMut<'a>: Iterator<Item = &'a mut dyn Any>
       where Self: 'static;

    const LEN: usize;

    // Required methods
    fn extend<Suffix>(self, suffix: Suffix) -> Self::Extend<Suffix>
       where Suffix: VariadicExt;
    fn reverse(self) -> Self::Reverse;
    fn as_ref_var(&self) -> Self::AsRefVar<'_>;
    fn as_mut_var(&mut self) -> Self::AsMutVar<'_>;
    fn iter_any_ref(&self) -> Self::IterAnyRef<'_>
       where Self: 'static;
    fn iter_any_mut(&mut self) -> Self::IterAnyMut<'_>
       where Self: 'static;
}

Extension methods/types for Variadics.

This is a sealed trait.

Required Associated Types

type Extend<Suffix>: VariadicExt where Suffix: VariadicExt

Creates a new (longer) variadic type by appending Suffix onto the end of this variadc.

type Reverse: VariadicExt

The reverse of this variadic type.

type AsRefVar<'a>: RefVariadic<UnRefVar = Self, RefVar = Self::AsRefVar<'a>, MutVar = Self::AsMutVar<'a>> where Self: 'a

This as a variadic of references.

type AsMutVar<'a>: MutVariadic<UnRefVar = Self, RefVar = Self::AsRefVar<'a>, MutVar = Self::AsMutVar<'a>> where Self: 'a

This as a variadic of exclusive (mut) references.

type IterAnyRef<'a>: Iterator<Item = &'a dyn Any> where Self: 'static

Iterator type returned by Self::iter_any_ref.

type IterAnyMut<'a>: Iterator<Item = &'a mut dyn Any> where Self: 'static

Iterator type returned by Self::iter_any_mut.

Required Associated Constants

const LEN: usize

The number of items in this variadic (its length).

Required Methods

fn extend<Suffix>(self, suffix: Suffix) -> Self::Extend<Suffix>
where Suffix: VariadicExt,

Extends this variadic value by appending suffix onto the end.

fn reverse(self) -> Self::Reverse

Reverses this variadic value.

fn as_ref_var(&self) -> Self::AsRefVar<'_>

Convert a reference to this variadic into a variadic of references.

let as_ref: var_type!(&u32, &String, &bool) =
    var_expr!(1_u32, "Hello".to_owned(), false).as_ref_var();

fn as_mut_var(&mut self) -> Self::AsMutVar<'_>

Convert an exclusive (mut) reference to this variadic into a variadic of exclusive (mut) references.

let as_mut: var_type!(&mut u32, &mut String, &mut bool) =
    var_expr!(1_u32, "Hello".to_owned(), false).as_mut_var();

fn iter_any_ref(&self) -> Self::IterAnyRef<'_>
where Self: 'static,

Iterate this variadic as &dyn Any references.

fn iter_any_mut(&mut self) -> Self::IterAnyMut<'_>
where Self: 'static,

Iterate this variadic as &mut dyn Any exclusive references.

Object Safety

This trait is not object safe.