variadics

Trait Split

pub trait Split<Prefix>: VariadicExt + Sealed<Prefix>
where
    Prefix: VariadicExt,
{
    type Suffix: VariadicExt;

    // Required methods
    fn split(self) -> (Prefix, Self::Suffix);
    fn split_ref(
        this: Self::AsRefVar<'_>,
    ) -> (Prefix::AsRefVar<'_>, <Self::Suffix as VariadicExt>::AsRefVar<'_>);
}

Helper trait for splitting a variadic into two parts. Prefix is the first part, everything after is the Suffix or second part.

This is a sealed trait.

Required Associated Types

type Suffix: VariadicExt

The second part when splitting this variadic by Prefix.

Required Methods

fn split(self) -> (Prefix, Self::Suffix)

Splits this variadic into two parts, first the Prefix, and second the Suffix.

fn split_ref( this: Self::AsRefVar<'_>, ) -> (Prefix::AsRefVar<'_>, <Self::Suffix as VariadicExt>::AsRefVar<'_>)

Splits a refvar variadic

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.