Trait variadics::Split

pub trait Split<Prefix>: Variadic + Sealed<Prefix>
where
    Prefix: Variadic,
{
    type Suffix: Variadic;

    // Required method
    fn split(self) -> (Prefix, Self::Suffix);
}

Helper trait for splitting a variadic into two parts. Prefix is the first part, everything after is the Suffix or second part.

This is a sealed trait.

Required Associated Types

type Suffix: Variadic

The second part when splitting this variadic by Prefix.

Required Methods

fn split(self) -> (Prefix, Self::Suffix)

Splits this variadic into two parts, first the Prefix, and second the Suffix.